amend! ldap: implement sasl authentication fallback
ldap: add support for sasl authentication

SASL implementation was implicitly disabled, which means that identities
were solely verified through the connection's given user and password.
This commit adds the ability to configure the authentication method and
adds a SASL authentication method option.

SASL-based authentication allows the LDAP bind to re-use the TLS
connection to provide the identity of the user.